django/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
psycopg==3.2.6
psycopg-binary==3.2.6
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
